Allow building and running on OSX
In addition to adding the correct natives OSX requires two
specific changes:

First we need to add GLFW forward compatability to allow OSX
to create a window. For more information see here:
http://www.glfw.org/faq.html#how-do-i-create-an-opengl-30-context

We also need to add `-XstartOnFirstThread` to prevent an error about
GLFW not being started on the first thread. This might be caused by
the fact that we're using AWT for image/texture loading.

However `-XstartOnFirstThread` makes `ImageIO.read` freeze, apparently
this is because of a deadlock which we can resolve by adding
`-Djava.awt.headless=true`.

Hopefully in the future we can fix this without a bunch of JVM flags
but for now it's fine.